fix: Updated file layout

+++ src/RoadRegistry.Product.Projections/GradeSeparatedJunctionRecordProjection.cs
@@ -16,15 +16,9 @@ namespace RoadRegistry.Product.Projections
     {
         public GradeSeparatedJunctionRecordProjection(RecyclableMemoryStreamManager manager, Encoding encoding)
         {
-            if (manager == null)
-            {
-                throw new ArgumentNullException(nameof(manager));
-            }
+            if (manager == null) throw new ArgumentNullException(nameof(manager));
 
-            if (encoding == null)
-            {
-                throw new ArgumentNullException(nameof(encoding));
-            }
+            if (encoding == null) throw new ArgumentNullException(nameof(encoding));
 
             When<Envelope<ImportedGradeSeparatedJunction>>(async (context, envelope, token) =>
             {
@@ -34,14 +28,14 @@ namespace RoadRegistry.Product.Projections
                     Id = envelope.Message.Id,
                     DbaseRecord = new GradeSeparatedJunctionDbaseRecord
                     {
-                        OK_OIDN = {Value = envelope.Message.Id},
-                        TYPE = {Value = translation.Identifier},
-                        LBLTYPE = {Value = translation.Name},
-                        BO_WS_OIDN = {Value = envelope.Message.UpperRoadSegmentId},
-                        ON_WS_OIDN = {Value = envelope.Message.LowerRoadSegmentId},
-                        BEGINTIJD = {Value = envelope.Message.Origin.Since},
-                        BEGINORG = {Value = envelope.Message.Origin.OrganizationId},
-                        LBLBGNORG = {Value = envelope.Message.Origin.Organization}
+                        OK_OIDN = { Value = envelope.Message.Id },
+                        TYPE = { Value = translation.Identifier },
+                        LBLTYPE = { Value = translation.Name },
+                        BO_WS_OIDN = { Value = envelope.Message.UpperRoadSegmentId },
+                        ON_WS_OIDN = { Value = envelope.Message.LowerRoadSegmentId },
+                        BEGINTIJD = { Value = envelope.Message.Origin.Since },
+                        BEGINORG = { Value = envelope.Message.Origin.OrganizationId },
+                        LBLBGNORG = { Value = envelope.Message.Origin.Organization }
                     }.ToBytes(manager, encoding)
                 };
 
@@ -51,7 +45,6 @@ namespace RoadRegistry.Product.Projections
             When<Envelope<RoadNetworkChangesAccepted>>(async (context, envelope, token) =>
             {
                 foreach (var change in envelope.Message.Changes.Flatten())
-                {
                     switch (change)
                     {
                         case GradeSeparatedJunctionAdded junctionAdded:
@@ -64,7 +57,6 @@ namespace RoadRegistry.Product.Projections
                             await RemoveJunction(context, junctionRemoved);
                             break;
                     }
-                }
             });
         }
 
@@ -81,17 +73,17 @@ namespace RoadRegistry.Product.Projections
                 Id = junction.Id,
                 DbaseRecord = new GradeSeparatedJunctionDbaseRecord
                 {
-                    OK_OIDN = {Value = junction.Id},
-                    TYPE = {Value = translation.Identifier},
-                    LBLTYPE = {Value = translation.Name},
-                    BO_WS_OIDN = {Value = junction.UpperRoadSegmentId},
-                    ON_WS_OIDN = {Value = junction.LowerRoadSegmentId},
+                    OK_OIDN = { Value = junction.Id },
+                    TYPE = { Value = translation.Identifier },
+                    LBLTYPE = { Value = translation.Name },
+                    BO_WS_OIDN = { Value = junction.UpperRoadSegmentId },
+                    ON_WS_OIDN = { Value = junction.LowerRoadSegmentId },
                     BEGINTIJD =
                     {
                         Value = LocalDateTimeTranslator.TranslateFromWhen(envelope.Message.When)
                     },
-                    BEGINORG = {Value = envelope.Message.OrganizationId},
-                    LBLBGNORG = {Value = envelope.Message.Organization}
+                    BEGINORG = { Value = envelope.Message.OrganizationId },
+                    LBLBGNORG = { Value = envelope.Message.Organization }
                 }.ToBytes(manager, encoding)
             };
 
@@ -108,7 +100,6 @@ namespace RoadRegistry.Product.Projections
 
             var translation = GradeSeparatedJunctionType.Parse(junction.Type).Translation;
             if (junctionRecord != null)
-            {
                 junctionRecord.DbaseRecord = new GradeSeparatedJunctionDbaseRecord
                 {
                     OK_OIDN = { Value = junction.Id },
@@ -120,17 +111,13 @@ namespace RoadRegistry.Product.Projections
                     BEGINORG = { Value = envelope.Message.OrganizationId },
                     LBLBGNORG = { Value = envelope.Message.Organization }
                 }.ToBytes(manager, encoding);
-            }
         }
 
         private static async Task RemoveJunction(ProductContext context, GradeSeparatedJunctionRemoved junction)
         {
             var junctionRecord = await context.GradeSeparatedJunctions.FindAsync(junction.Id);
 
-            if (junctionRecord != null)
-            {
-                context.GradeSeparatedJunctions.Remove(junctionRecord);
-            }
+            if (junctionRecord != null) context.GradeSeparatedJunctions.Remove(junctionRecord);
         }
     }
 }
